requirements.txt
requests==2.31.0
numpy>=1.24.0,<1.26.0
pandas>=2.0.0,<2.2.0
feedgen==1.0.0
python-dotenv==1.0.0
pytz==2024.1
beautifulsoup4==4.12.0
notion-client==2.2.1
